Move pick function to Camera
